iot-simulation/requirements.txt
firebase-admin==6.4.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
schedule==1.2.0
scikit-learn==1.3.0
flask==2.3.3
flask-cors==4.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
